Separating RuntimeCache API from implementation
