+++ engine/runtime-instrument-common/src/main/java/org/enso/interpreter/instrument/RuntimeCache.java
@@ -1,11 +1,6 @@
 package org.enso.interpreter.instrument;
 
-import com.oracle.truffle.api.CompilerDirectives;
-import java.lang.ref.Reference;
-import java.lang.ref.SoftReference;
-import java.lang.ref.WeakReference;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
 import java.util.function.Consumer;
@@ -13,199 +8,56 @@ import java.util.function.Supplier;
 import org.enso.common.CachePreferences;
 import org.enso.interpreter.service.ExecutionService;
 
-/** A storage for computed values. */
-public final class RuntimeCache implements java.util.function.Function<String, Object> {
-  private final Map<UUID, Reference<Object>> cache = new HashMap<>();
-  private final Map<UUID, Reference<Object>> expressions = new HashMap<>();
-  private final Map<UUID, TypeInfo> types = new HashMap<>();
-  private final Map<UUID, ExecutionService.FunctionCallInfo> calls = new HashMap<>();
-  private CachePreferences preferences = CachePreferences.empty();
-  private Consumer<UUID> observer;
+/**
+ * Immutable API facade for Enso runtime cache. It contains only "query" methods. See {@link
+ * Mutable} for operations that can modify the cache.
+ *
+ * <p>Implementation is provided by original implementation in {@link RuntimeCacheImpl}.
+ */
+public abstract class RuntimeCache {
+  /** the only implementation is in the same package */
+  RuntimeCache() {}
+
+  /** Factory method to create new runtime cache. */
+  public static RuntimeCache create() {
+    return new RuntimeCacheImpl();
+  }
 
   /**
-   * Add value to the cache if it is possible.
+   * Reads a cached value from the cache for given key. Cached values are hold be "soft reference" -
+   * e.g. they are kept until the system runs out of memory or until the value for given key is
+   * replacd
    *
-   * @param key the key of an entry.
-   * @param value the added value.
-   * @return {@code true} if the value was added to the cache.
+   * @param key the UUID of the key
+   * @return cached value associated with the {@code key} or {@code null}
    */
-  @CompilerDirectives.TruffleBoundary
-  public boolean offer(UUID key, Object value) {
-    expressions.put(key, new WeakReference<>(value));
-    if (preferences.contains(key)) {
-      var ref = new SoftReference<>(value);
-      cache.put(key, ref);
-      return true;
-    }
-    return false;
-  }
-
-  /** Get the value from the cache. */
-  public Object get(UUID key) {
-    var ref = cache.get(key);
-    var res = ref != null ? ref.get() : null;
-    return res;
-  }
-
-  /** Get the value from the cache. */
-  public Object getAnyValue(UUID key) {
-    var ref = expressions.get(key);
-    var res = ref != null ? ref.get() : null;
-    return res;
-  }
-
-  // Accessed in InstrumentorBuiltin
-  @Override
-  public Object apply(String uuid) {
-    Object res;
-    try {
-      var key = UUID.fromString(uuid);
-      var ref = expressions.get(key);
-      res = ref != null ? ref.get() : null;
-      var callback = observer;
-      if (callback != null) {
-        callback.accept(key);
-      }
-    } catch (IllegalArgumentException ex) {
-      res = null;
-    }
-    return res;
-  }
-
-  /** Remove the value from the cache. */
-  public Object remove(UUID key) {
-    var ref = cache.remove(key);
-    return ref == null ? null : ref.get();
-  }
+  public abstract Object get(UUID key);
 
   /**
-   * @return all cache keys.
-   */
-  public Set<UUID> getKeys() {
-    return cache.keySet();
-  }
-
-  /** Clear the cached values. */
-  public void clear() {
-    cache.clear();
-  }
-
-  /**
-   * Clear cached values of the provided kind.
+   * Reads a value that was notified to the cache (which may or may not be subject to caching). A
+   * non-cached values include values of expression which are only available while the computation
+   * is running.
    *
-   * @param kind the kind of cached value to clear
-   * @return the set of cleared keys
+   * @param key the UUID of the key
+   * @return available value or {@code null}
    */
-  public Set<UUID> clear(CachePreferences.Kind kind) {
-    var keys = preferences.get(kind);
-    for (var key : keys) {
-      cache.remove(key);
-    }
-    return keys;
-  }
+  public abstract Object getAnyValue(UUID key);
 
   /**
-   * Cache the type of expression.
+   * Obtains info about a type of expression identify by UUID.
    *
-   * @return the previously cached type.
+   * @param key the UUID of the key
+   * @return the cached type of the expression or {@code null}
    */
-  @CompilerDirectives.TruffleBoundary
-  public TypeInfo putType(UUID key, TypeInfo typeInfo) {
-    return types.put(key, typeInfo);
-  }
+  public abstract TypeInfo getType(UUID key);
 
   /**
-   * @return the cached type of the expression
-   */
-  @CompilerDirectives.TruffleBoundary
-  public TypeInfo getType(UUID key) {
-    return types.get(key);
-  }
-
-  /**
-   * Cache the function call
+   * Obtains function call for given UUID.
    *
-   * @param key the expression associated with the function call.
-   * @param call the function call.
-   * @return the function call that was previously associated with this expression.
+   * @param key the UUID of the key
+   * @return the cached function call associated with the expression or {@code null}
    */
-  @CompilerDirectives.TruffleBoundary
-  public ExecutionService.FunctionCallInfo putCall(
-      UUID key, ExecutionService.FunctionCallInfo call) {
-    if (call == null) {
-      return calls.remove(key);
-    }
-    return calls.put(key, call);
-  }
-
-  /**
-   * @return the cached function call associated with the expression.
-   */
-  @CompilerDirectives.TruffleBoundary
-  public ExecutionService.FunctionCallInfo getCall(UUID key) {
-    return calls.get(key);
-  }
-
-  /**
-   * @return the cached method calls.
-   */
-  public Set<UUID> getCalls() {
-    return calls.keySet();
-  }
-
-  /**
-   * Remove the function call from the cache.
-   *
-   * @param key the expression associated with the function call.
-   */
-  public void removeCall(UUID key) {
-    calls.remove(key);
-  }
-
-  /** Clear the cached calls. */
-  public void clearCalls() {
-    calls.clear();
-  }
-
-  /** Remove the type associated with the provided key. */
-  public void removeType(UUID key) {
-    types.remove(key);
-  }
-
-  /** Clear the cached types. */
-  public void clearTypes() {
-    types.clear();
-  }
-
-  /**
-   * @return the preferences of this cache.
-   */
-  public CachePreferences getPreferences() {
-    return preferences;
-  }
-
-  /**
-   * Set the new cache preferences.
-   *
-   * @param preferences the new cache preferences
-   */
-  public void setPreferences(CachePreferences preferences) {
-    this.preferences = preferences;
-  }
-
-  /**
-   * Remove the cache preference associated with the provided key.
-   *
-   * @param key the preference to remove
-   */
-  public void removePreference(UUID key) {
-    preferences.remove(key);
-  }
-
-  /** Clear the cache preferences. */
-  public void clearPreferences() {
-    preferences.clear();
-  }
+  public abstract ExecutionService.FunctionCallInfo getCall(UUID key);
 
   /**
    * Executes a query while tracking access to the cache by {@code callback} observer.
@@ -215,13 +67,66 @@ public final class RuntimeCache implements java.util.function.Function<String, O
    * @return value computed by the {@code scope}
    * @param <V> type of the returned value
    */
-  public <V> V runQuery(Consumer<UUID> callback, Supplier<V> scope) {
-    var previousCallback = this.observer;
-    this.observer = callback;
-    try {
-      return scope.get();
-    } finally {
-      this.observer = previousCallback;
+  public abstract <V> V runQuery(Consumer<UUID> callback, Supplier<V> scope);
+
+  /**
+   * Obtains a list of UUIDs known to the cache.
+   *
+   * @param calls collect UUIDs of the calls
+   * @param preferences collect UUIDs of the preferences
+   * @return
+   */
+  public final Set<UUID> findUUIDs(boolean calls, boolean preferences) {
+    var impl = (RuntimeCacheImpl) this;
+    var set = new HashSet<UUID>();
+    if (calls) {
+      set.addAll(impl.getCalls());
     }
+    if (preferences) {
+      set.addAll(impl.getPreferences().preferences().keySet());
+    }
+    return set;
+  }
+
+  /**
+   * Checks whether this key is associated with a binding expression.
+   *
+   * @param uuid the key to check
+   * @return {@code true} or {@code false}
+   */
+  public final boolean isBindingExpression(UUID uuid) {
+    var impl = (RuntimeCacheImpl) this;
+    return impl.getPreferences().get(uuid) == CachePreferences.Kind.BINDING_EXPRESSION;
+  }
+
+  /** */
+  public interface Mutable {
+    public abstract void clear();
+
+    /**
+     * Add value to the cache if it is possible.
+     *
+     * @param key the key of an entry.
+     * @param value the added value.
+     * @return {@code true} if the value was added to the cache.
+     */
+    public boolean offer(UUID key, Object value);
+
+    /**
+     * Cache the type of expression.
+     *
+     * @return the previously cached type.
+     */
+    public TypeInfo putType(UUID key, TypeInfo typeInfo);
+
+    /**
+     * Cache the function call
+     *
+     * @param key the expression associated with the function call.
+     * @param call the function call.
+     * @return the function call that was previously associated with this expression.
+     */
+    public ExecutionService.FunctionCallInfo putCall(
+        UUID key, ExecutionService.FunctionCallInfo call);
   }
 }
